refactoring, basic support for array types, WIP

+++ test/support/clear_redis.rb
@@ -1,2 +1,2 @@
 require 'redis'
-$redis.flushall
+$redis.flushall if defined?($redis) and not $redis.nil?
